refactor(user): remove unecessary list state
