+++ packages/manager/modules/cda/src/details/cda-details.html
@@ -12,7 +12,7 @@
                 <span data-translate="cda_tab_home"></span>
             </oui-header-tabs-item>
             <oui-header-tabs-item
-                state="cda.dashboard.cda-user.cda-user-list"
+                state="cda.dashboard.cda-user"
                 state-params="{ serviceName: CdaDetailsCtrl.serviceName }"
             >
                 <span data-translate="cda_tab_user"></span>
