+++ packages/manager/modules/cda/src/user/index.js
@@ -1,7 +1,7 @@
 import angular from 'angular';
 
-import routing from './cda-user';
-import userList from './list';
+import routing from './routing';
+
 import userDetails from './details';
 import addController from './add/cda-user-add.controller';
 import deleteController from './delete/cda-user-delete.controller';
@@ -9,7 +9,7 @@ import deleteController from './delete/cda-user-delete.controller';
 const moduleName = 'ovhManagerCdaUser';
 
 angular
-  .module(moduleName, [userDetails, userList])
+  .module(moduleName, [userDetails])
   .config(routing)
   .controller('CdaUserAddCtrl', addController)
   .controller('CdaUserDeleteCtrl', deleteController)
