--- packages/manager/modules/cda/src/user/list/cda-user-list.html
@@ -1,56 +1,0 @@
-<section>
-    <header>
-        <p data-translate="cda_user_permission_definition"></p>
-    </header>
-    <div>
-        <button
-            type="button"
-            class="oui-button oui-button_secondary"
-            data-ng-click="CdaUserListCtrl.openAddModal()"
-            data-ng-if="!CdaUserListCtrl.loading"
-        >
-            <span data-translate="cda_user_add_button_text"></span>
-        </button>
-    </div>
-    <oui-datagrid data-rows="CdaUserListCtrl.datas.users">
-        <oui-datagrid-column
-            data-title="'cda_user_table_col_username' | translate"
-            data-property="user.name"
-            data-sortable="asc"
-        ></oui-datagrid-column>
-        <oui-datagrid-column
-            data-title="'cda_user_table_col_key' | translate"
-            data-property="user.key"
-            data-sortable
-        ></oui-datagrid-column>
-        <oui-datagrid-column
-            data-title="'cda_user_table_col_pool' | translate"
-            data-property="pools"
-            data-sortable
-        >
-            <span
-                data-ng-repeat="pool in $row.pools | limitTo: CdaUserListCtrl.options.maxPoolDisplay track by pool.poolName"
-            >
-                {{ pool.poolName }}{{ !$last ||
-                CdaUserListCtrl.isTruncatedPoolArray($row.pools , $index) ? ','
-                : '' }}
-                <span
-                    data-ng-if="$last && CdaUserListCtrl.isTruncatedPoolArray($row.pools , $index)"
-                    >...</span
-                >
-            </span>
-        </oui-datagrid-column>
-
-        <oui-action-menu data-compact data-placement="end">
-            <oui-action-menu-item
-                data-on-click="CdaUserListCtrl.viewPermissions($row.user.name)"
-                ><span data-translate="cda_user_edit_button_text"></span
-            ></oui-action-menu-item>
-            <oui-action-menu-divider></oui-action-menu-divider>
-            <oui-action-menu-item
-                data-on-click="CdaUserListCtrl.openDeleteModal($row)"
-                ><span data-translate="cda_user_delete_button_text"></span
-            ></oui-action-menu-item>
-        </oui-action-menu>
-    </oui-datagrid>
-</section>
