--- packages/manager/modules/cda/src/user/list/cda-user-list.js
@@ -1,19 +1,0 @@
-import controller from './cda-user-list.controller';
-import template from './cda-user-list.html';
-
-export default /* @ngInject */ ($stateProvider) => {
-  $stateProvider.state('cda.dashboard.cda-user.cda-user-list', {
-    url: '/list',
-    views: {
-      cdaUserContent: {
-        template,
-        controller,
-        controllerAs: 'CdaUserListCtrl',
-      },
-    },
-    translations: {
-      format: 'json',
-      value: ['.'],
-    },
-  });
-};
