--- packages/manager/modules/cda/src/user/list/index.js
@@ -1,12 +1,0 @@
-import angular from 'angular';
-
-import routing from './cda-user-list';
-
-const moduleName = 'ovhManagerCdaUserList';
-
-angular
-  .module(moduleName, [])
-  .config(routing)
-  .run(/* @ngTranslationsInject:json ./translations */);
-
-export default moduleName;
