--- packages/manager/modules/cda/src/user/list/translations/Messages_en_GB.json
@@ -1,21 +1,0 @@
-{
-  "cda_user_permission_definition": "Your users can use the cluster with pool-specific permissions.",
-  "cda_user_add_button_text": "Add user",
-  "cda_user_delete_button_text": "Delete",
-  "cda_user_edit_button_text": "See permissions",
-  "cda_user_table_col_username": "Name",
-  "cda_user_table_col_key": "Key",
-  "cda_user_table_col_pool": "Pools",
-  "cda_user_table_col_actions": "Actions",
-  "cda_user_add_title": "Add user",
-  "cda_user_add_username_label": "Name",
-  "cda_user_add_username_required": "You must name your cluster.",
-  "cda_user_add_username_pattern": "The name cannot contain spaces.",
-  "cda_user_add_username_maxlength": "The name of your cluster must not exceed {{ maxlength }} characters.",
-  "cda_user_add_success": "The user has been successfully added.",
-  "cda_user_delete_title": "Delete user",
-  "cda_user_delete_username_label": "Name:",
-  "cda_user_delete_confirm": "Are you sure you want to delete this user?",
-  "cda_user_delete_success": "The user has been successfully deleted.",
-  "cda_user_table_empty": "No user"
-}
