--- packages/manager/modules/cda/src/user/list/translations/Messages_pt_PT.json
@@ -1,21 +1,0 @@
-{
-  "cda_user_permission_definition": "Os seus utilizadores podem utilizar o cluster com permissões específicas por pool.",
-  "cda_user_add_button_text": "Adicionar utilizador",
-  "cda_user_delete_button_text": "Eliminar",
-  "cda_user_edit_button_text": "Ver as permissões",
-  "cda_user_table_col_username": "Apelido",
-  "cda_user_table_col_key": "Chave",
-  "cda_user_table_col_pool": "Pools",
-  "cda_user_table_col_actions": "Ações",
-  "cda_user_add_title": "Adicionar utilizador",
-  "cda_user_add_username_label": "Apelido",
-  "cda_user_add_username_required": "O nome é obrigatório.",
-  "cda_user_add_username_pattern": "O nome não pode conter espaços.",
-  "cda_user_add_username_maxlength": "O nome deve ter um comprimento máximo de {{ maxlength }} caracteres.",
-  "cda_user_add_success": "Utilizador adicionado com sucesso.",
-  "cda_user_delete_title": "Eliminar utilizador ",
-  "cda_user_delete_username_label": "Nome",
-  "cda_user_delete_confirm": "Tem a certeza de que deseja eliminar este utilizador?",
-  "cda_user_delete_success": "Utilizador eliminado com sucesso.",
-  "cda_user_table_empty": "Nenhum utilizador"
-}
